+++ balance.mjs
@@ -2,4 +2,4 @@
 export * from './src/index.mjs';
 import { cli } from './src/index.mjs';
 
-await cli({ moduleUrl: import.meta.url });
+await cli({ moduleUrl: import.meta.url, argv: process.argv.slice(2) });